Implement a new mechanism for obtaining a socket's peer credentials
The snapd mechanism relies on encoding the credentials to a string,
storing it as the connection's peer address and then using a regular
expression to decode the string.

This avoids the string conversion by configuring http.Server to create
a per-connection context and then attaching the connections's net.Conn
to the context. As the http package then uses the connection context
to create the http.Request, it is possible to obtain the connection in
the command handler and then use this to query the kernel for the peer
credentials.